Added Anisotropy Script
This script switches between VV and VH several times at a fixed
wavelengh in order to calculate anisotropy and control fluorophore
degradation.
The G factor must be calculated in a separate measurement.
The script can be used with CW or pulsed excitation.  In the latter
case the decays can also be used to calculate time resolved anisotropy